+++ cluster.go
@@ -81,30 +81,36 @@ func New(points []GeoPoint, opts ...Option) (*Cluster, error) {
 // returns the array of clustered points,
 // X coordinate of returned object is Longitude and Y coordinate of returned object is Latitude.
 func (c *Cluster) GetClusters(northWest, southEast GeoPoint, zoom int, limit int) []Point {
+	nw := northWest.GetCoordinates()
+	se := southEast.GetCoordinates()
+
+	if nw == nil || se == nil {
+		return nil
+	}
 	// Original mapbox/supercluster library code has the following expression to calculate min and max longitudes:
 	// let minLng = ((bbox[0] + 180) % 360 + 360) % 360 - 180;
 	// Mozilla developer guide suggests such construction to obtain a modulo
 	// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Remainder:
 	// > To obtain a modulo in JavaScript, in place of a % n, use ((a % n ) + n ) % n.
-	minLng := math.Mod(math.Mod(southEast.GetCoordinates().Lng+180, 360)+360, 360) - 180
-	minLat := math.Max(-90, math.Min(90, southEast.GetCoordinates().Lat))
+	minLng := math.Mod(math.Mod(nw.Lng+180, 360)+360, 360) - 180
+	minLat := math.Max(-90, math.Min(90, se.Lat))
 
 	var maxLng float64
 
-	if northWest.GetCoordinates().Lng != 180 {
-		maxLng = math.Mod(math.Mod(northWest.GetCoordinates().Lng+180, 360)+360, 360) - 180
-	} else {
+	if se.Lng == 180 {
 		maxLng = 180
+	} else {
+		maxLng = math.Mod(math.Mod(se.Lng+180, 360)+360, 360) - 180
 	}
 
-	maxLat := math.Max(-90, math.Min(90, northWest.GetCoordinates().Lat))
+	maxLat := math.Max(-90, math.Min(90, nw.Lat))
 
-	if northWest.GetCoordinates().Lng-southEast.GetCoordinates().Lng >= 360 {
+	if se.Lng-nw.Lng >= 360 {
 		minLng = -180
 		maxLng = 180
 	} else if minLng > maxLng {
-		easternHem := c.GetClusters(&Point{X: 180, Y: maxLat}, &Point{X: minLng, Y: minLat}, zoom, limit)
-		westernHem := c.GetClusters(&Point{X: maxLng, Y: maxLat}, &Point{X: -180, Y: minLat}, zoom, limit)
+		easternHem := c.GetClusters(&Point{X: minLng, Y: maxLat}, &Point{X: 180, Y: minLat}, zoom, limit)
+		westernHem := c.GetClusters(&Point{X: -180, Y: maxLat}, &Point{X: maxLng, Y: minLat}, zoom, limit)
 
 		return append(easternHem, westernHem...)
 	}